requirements.txt
asgiref>=3.7.2
bidict>=0.22.1
Django>=4.2.3
django-ninja>=0.22.2
Pillow>=10.0.0
pydantic>=1.10.11
python-engineio>=4.5.1
python-socketio>=5.8.0
sqlparse>=0.4.4
typing_extensions==4.7.1
